+++ src/pkgIndex.tcl
@@ -1,4 +1,4 @@
-package ifneeded alited 1.8.8.4 [list source [file join $dir alited.tcl]]
+package ifneeded alited 1.8.8.5 [list source [file join $dir alited.tcl]]
 
 namespace eval ::alited {